Get BuyOrderState one big connected component, and let user view failure
