--- packages/instant/src/components/retry_button.tsx
@@ -1,11 +1,0 @@
-import * as React from 'react';
-
-import { SecondaryButton, SecondaryButtonProps } from './secondary_button';
-
-export interface RetryButtonProps extends SecondaryButtonProps {
-    onClick: () => void;
-}
-
-export const RetryButton: React.StatelessComponent<RetryButtonProps> = props => {
-    return <SecondaryButton {...props}>Try Again</SecondaryButton>;
-};
